% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures_old.R
\name{gausskld}
\alias{gausskld}
\title{gausskld}
\usage{
gausskld(P)
}
\arguments{
\item{P}{dataframe (2 rows only) containing gaussian params. Colnames should be mu and sd}
}
\value{
kl divergence
}
\description{
kl divergence between 2 gaussians
}
