% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpuApi.R
\name{hashesFromImages}
\alias{hashesFromImages}
\title{hashesFromImages}
\usage{
curl -v http://localhost:8004/ocpu/library/whaleRidgeFindR/R/hashesFromImages/json\
-F "img1=@C:/Users/jathompson/Documents/Testingdb/jensImgs/test2.jpg"\
-F "img2=@C:/Users/jathompson/Documents/Testingdb/jensImgs/test3.jpg"\
-F "test=@C:/Users/jathompson/Documents/Testingdb/jensImgs/test4.jpg"\
-F "misc=@C:/Users/jathompson/Documents/Testingdb/jensImgs/test7.jpg"
}
\arguments{
\item{variable}{number of character vectors which denotes image file "directory/whaleRidgeImage.JPG" to be processed in parallel}
}
\value{
list of two sub lists respectively containing:
"hash" vector specifying an individual
"coordinates" a 2 column matrix of coordinates
each element in this list is nammed with the argument passed in via the api
}
\description{
hashesFromImages
}
\details{
\code{traceFromImage} wrapper for use through opencpu.
opencpu passes temp object name to function followed by \code{traceToHash}
Processes an image(cimg) containing a whaleRidge. 
First the image undergoes cleanup through a variety of filters and glare removal via
\code{constrainSizeFinImage} and \code{fillGlare}
These processes help enhance edge clarity.
The trailing edge is highlighted via neural network. 
The image is then cropped down to the trailing edge for efficiency purposes.
The canny edges are then extracted from the crop and passed to 
\code{traceFromCannyEdges}
which isolates coordinates for the trailing edge. These coordinates are then passed to
\code{extractAnnulus}
which collects image data used for identification.
Both the coordinates and the image annulus are then returned.
}
