% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures_old.R
\name{knn}
\alias{knn}
\title{knn}
\usage{
knn(X, k)
}
\arguments{
\item{X}{vector to fit}

\item{k}{number of distributions}
}
\value{
list of distribution paramiters, assignments and density
}
\description{
rough fit of gaussians via knn
}
