% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures_old.R
\name{pClip}
\alias{pClip}
\title{pClip}
\usage{
pClip(x, a, b)
}
\arguments{
\item{x}{is value to clip}

\item{a}{is min}

\item{b}{is max}
}
\value{
clipped x
}
\description{
slopy version of clamp
}
