% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures_old.R
\name{shrinkDomDim}
\alias{shrinkDomDim}
\title{shrinkDomDim}
\usage{
shrinkDomDim(image, maxDim)
}
\arguments{
\item{image}{image to resize}

\item{maxDim}{maximum dim}
}
\value{
resizeed image
}
\description{
shrinkDomDim
}
\details{
shrink largest dim to maxDim in ratio preserving way
}
